Pythonic refactor: io_util.py -> ioutil.py
One of several refactors to remove the use of utility classes which only
contain static methods.
