+++ pd_utils/cli/coverage_gap_report_cli.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 
 from pd_utils.report import CoverageGapReport
 from pd_utils.util import datetool
-from pd_utils.util import IOUtil
+from pd_utils.util import ioutil
 from pd_utils.util import RuntimeInit
 
 
@@ -25,8 +25,8 @@ def main(*, _args: list[str] | None = None) -> int:
     schedule_report, escalation_report = client.run_reports()
 
     now = datetool.utcnow_isotime().split("T")[0]
-    IOUtil.write_to_file(f"schedule_gap_report{now}.csv", schedule_report)
-    IOUtil.write_to_file(f"escalation_rule_gap_report{now}.csv", escalation_report)
+    ioutil.write_to_file(f"schedule_gap_report{now}.csv", schedule_report)
+    ioutil.write_to_file(f"escalation_rule_gap_report{now}.csv", escalation_report)
 
     return 0
