+++ pd_utils/cli/user_report_cli.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 
 from pd_utils.report import UserReport
 from pd_utils.util import datetool
-from pd_utils.util import IOUtil
+from pd_utils.util import ioutil
 from pd_utils.util import RuntimeInit
 
 
@@ -27,7 +27,7 @@ def main(_args: list[str] | None = None) -> int:
     report = UserReport(args.token).run_report(team_ids=args.team_ids)
 
     now = datetool.utcnow_isotime().split("T")[0]
-    IOUtil.write_to_file(f"user_report{now}.csv", report)
+    ioutil.write_to_file(f"user_report{now}.csv", report)
     print(f"Report saved to user_report{now}.csv")
 
     return 0
