+++ pd_utils/util/__init__.py
@@ -1,11 +1,9 @@
 from __future__ import annotations
 
-from .io_util import IOUtil
 from .pagerduty_api import PagerDutyAPI
 from .runtime_init import RuntimeInit
 
 __all__ = [
-    "IOUtil",
     "PagerDutyAPI",
     "RuntimeInit",
 ]
